migrate `exhaustive_cases` integration test
