+++ test/integration_test.dart
@@ -19,7 +19,6 @@ import 'integration/always_require_non_null_named_parameters.dart'
 import 'integration/avoid_web_libraries_in_flutter.dart'
     as avoid_web_libraries_in_flutter;
 import 'integration/close_sinks.dart' as close_sinks;
-import 'integration/exhaustive_cases.dart' as exhaustive_cases;
 import 'integration/public_member_api_docs.dart' as public_member_api_docs;
 import 'integration/use_build_context_synchronously.dart'
     as use_build_context_synchronously;
