--- test_data/integration/exhaustive_cases/a.dart
@@ -1,16 +1,0 @@
-// Copyright (c) 2020, the Dart project authors. Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-import 'e.dart' as prefixed;
-
-void e(prefixed.E e) {
-  // Missing case.
-  switch(e) { // LINT
-    case prefixed.E.e :
-      print('e');
-      break;
-    case prefixed.E.f :
-      print('e');
-  }
-}
