--- test_data/integration/exhaustive_cases/e.dart
@@ -1,13 +1,0 @@
-// Copyright (c) 2020, the Dart project authors. Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-// Enum-like
-class E {
-  final int i;
-  const E._(this.i);
-
-  static const e = E._(1);
-  static const f = E._(2);
-  static const g = E._(3);
-}
